Refactor to make testable.  Fix code template for serpentine-ness.
